Renamed folder `interface` to `traits`
